pyproject.toml
[project]
name = "tfg-leyre"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymupdf==1.23.8",
    "langchain==0.1.4",
    "langchain-community==0.0.16",
    "psycopg2-binary==2.9.9",
    "pgvector==0.2.4",
    "sentence-transformers",
    "python-dotenv"
]
